pyproject.toml
[tool.black]
line-length = 120
skip-string-normalization = true
exclude = '''
    /(
        migrations
    )/
'''
[tool.poetry]
name = "finsey"
version = "0.1.0"
description = "A personal finance management tool"
authors = ["Agnieszka Wojdecka <agnieszka.wojdecka@wp.pl>"]
license = "MIT"
readme = "README.md"

[tool.isort]
profile = "black"
skip = ["migrations"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pygments = "^2.18.0"
celery = {extras = ["redis"], version = "^5.4.0"}
psycopg2-binary = "^2.9.10"
django-celery-beat = "^2.7.0"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.18.1"
flake8 = "^7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
